requirements.txt
langchain_openai~=0.2
langsmith~=0.1.139
langchain~=0.3.1
python-dotenv~=1.0
sqlalchemy~=2.0
tweepy~=4.14
twikit~=2.2
